# docs/swagger.yaml
info:
  contact: {}
paths:
  /login:
    post:
      parameters:
      - description: username
        in: formData
        name: username
        type: string
      - description: password
        in: formData
        name: password
        type: string
      responses:
        "200":
          description: '{"code":"200","msg","","data":""}'
          schema:
            type: string
      summary: 用户登录
      tags:
      - 公共方法
  /problem-detail:
    get:
      parameters:
      - description: problem identity
        in: query
        name: identity
        type: string
      responses:
        "200":
          description: '{"code":"200","msg","","data":""}'
          schema:
            type: string
      summary: 问题详情
      tags:
      - 公共方法
  /problem-list:
    get:
      parameters:
      - description: page
        in: query
        name: page
        type: integer
      - description: size
        in: query
        name: size
        type: integer
      - description: keyword
        in: query
        name: keyword
        type: string
      - description: category_identity
        in: query
        name: category_identity
        type: string
      responses:
        "200":
          description: '{"code":"200","msg","","data":""}'
          schema:
            type: string
      summary: 问题列表
      tags:
      - 公共方法
  /send-code:
    post:
      parameters:
      - description: email
        in: formData
        name: email
        required: true
        type: string
      responses:
        "200":
          description: '{"code":"200","msg","","data":""}'
          schema:
            type: string
      summary: 发送验证码
      tags:
      - 公共方法
  /submit-list:
    get:
      parameters:
      - description: page
        in: query
        name: page
        type: integer
      - description: size
        in: query
        name: size
        type: integer
      - description: problem_identity
        in: query
        name: problem_identity
        type: string
      - description: user_identity
        in: query
        name: user_identity
        type: string
      - description: status
        in: query
        name: status
        type: integer
      responses:
        "200":
          description: '{"code":"200","msg","","data":""}'
          schema:
            type: string
      summary: 提交列表
      tags:
      - 公共方法
  /user-detail:
    get:
      parameters:
      - description: user identity
        in: query
        name: identity
        type: string
      responses:
        "200":
          description: '{"code":"200","msg","","data":""}'
          schema:
            type: string
      summary: 用户详情
      tags:
      - 公共方法
swagger: "2.0"